Search parameters are moved to Share button component
