+++ src/components/FloatingToolBar/ShareNetworkButtton.tsx
@@ -1,21 +1,75 @@
 import { Alert, IconButton, Snackbar, Tooltip } from '@mui/material'
 import { Share } from '@mui/icons-material'
 import { useWorkspaceStore } from '../../store/WorkspaceStore'
-import { useState } from 'react'
+import { useEffect, useState } from 'react'
+import { useSearchParams } from 'react-router-dom'
+import { Ui } from '../../models/UiModel'
+import { NetworkView } from '../../models/ViewModel'
+import { useUiStateStore } from '../../store/UiStateStore'
+import { useViewModelStore } from '../../store/ViewModelStore'
+
+// Selection will be encoded if the selected object count is less than this number
+const MAX_SELECTED_OBJ = 300
 
 /**
  * Button to copy the sharable URL to clipboard
  *
- *
+ * This component watches the UI and Selection states and encode them as URL search params
  */
 export const ShareNetworkButton = (): JSX.Element => {
-  const [open, setOpen] = useState(false)
-
   // Use this as the staring point for the sharable URL
   const wsId = useWorkspaceStore((state) => state.workspace.id)
+
   const currentNetworkId = useWorkspaceStore(
     (state) => state.workspace.currentNetworkId,
   )
+
+  // Encode UI states as URL search params
+  const [search, setSearch] = useSearchParams()
+
+  const ui: Ui = useUiStateStore((state) => state.ui)
+  const { panels } = ui
+
+  // This will be used to watch the selection state
+  const networkViewModel: NetworkView = useViewModelStore(
+    (state) => state.viewModels[currentNetworkId],
+  )
+
+  const selectedNodeCount: number = networkViewModel?.selectedNodes.length
+  const selectedEdgeCount: number = networkViewModel?.selectedEdges.length
+
+  useEffect(() => {
+    // encode panel states as URL search params
+    console.log('Panels::', panels, search.keys())
+    setSearch({ ...search, ...panels })
+  }, [panels])
+
+  useEffect(() => {
+    console.log('Updating', networkViewModel)
+    if (networkViewModel === undefined) {
+      return
+    }
+
+    const params = new URLSearchParams(search)
+    const selectedNodeCount: number = networkViewModel.selectedNodes.length
+    const selectedEdgeCount: number = networkViewModel.selectedEdges.length
+    if (selectedNodeCount === 0 && selectedEdgeCount === 0) {
+      params.delete('selectednodes')
+      params.delete('selectededges')
+      setSearch(params)
+      return
+    }
+
+    if (selectedNodeCount > 0 && selectedNodeCount <= MAX_SELECTED_OBJ) {
+      params.set('selectednodes', networkViewModel.selectedNodes.join(' '))
+    }
+    if (selectedEdgeCount > 0 && selectedEdgeCount <= MAX_SELECTED_OBJ) {
+      params.set('selectededges', networkViewModel.selectedEdges.join(' '))
+    }
+    setSearch(params)
+  }, [selectedNodeCount, selectedEdgeCount])
+
+  const [open, setOpen] = useState(false)
 
   const copyTextToClipboard = async (text: string): Promise<void> => {
     if ('clipboard' in navigator) {
